# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/projects": {
      "get": {
        "operationId": "ProjectsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список проектов",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить все проекты текущего пользователя",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "ProjectsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Проект успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Slug уже занят"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Создать новый проект",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/{id}": {
      "get": {
        "operationId": "ProjectsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID проекта",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о проекте",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Проект не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить проект по ID",
        "tags": [
          "projects"
        ]
      },
      "put": {
        "operationId": "ProjectsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID проекта",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Проект обновлен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Проект не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Обновить проект",
        "tags": [
          "projects"
        ]
      },
      "delete": {
        "operationId": "ProjectsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID проекта",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Проект удален",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Проект не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удалить проект",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/{id}/media": {
      "post": {
        "operationId": "ProjectsController_addMedia",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID проекта",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Медиа файл добавлен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMediaResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Проект не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Добавить медиа файл к проекту",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/{id}/media/{mediaId}": {
      "delete": {
        "operationId": "ProjectsController_removeMedia",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID проекта",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "description": "ID медиа файла",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Медиа файл удален",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMediaResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Проект или медиа файл не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удалить медиа файл из проекта",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/{id}/ai/describe": {
      "post": {
        "operationId": "ProjectsController_generateDescription",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID проекта",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Описание проекта сгенерировано",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "description": {
                      "type": "string",
                      "example": "Современная платформа для электронной коммерции..."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Проект не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Сгенерировать описание проекта с помощью AI",
        "tags": [
          "projects"
        ]
      }
    },
...
  },
  "info": {
    "title": "PortfolioX API",
    "description": "Backend API для платформы PortfolioX - совместное резюме + портфолио",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ProjectMediaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID медиа файла",
            "example": "clx1234567890"
          },
          "projectId": {
            "type": "string",
            "description": "ID проекта",
            "example": "clx1234567890"
          },
          "url": {
            "type": "string",
            "description": "URL медиа файла",
            "example": "https://example.com/image.jpg"
          },
          "type": {
            "type": "string",
            "description": "Тип медиа файла",
            "enum": [
              "IMAGE",
              "VIDEO",
              "FILE"
            ],
            "example": "IMAGE"
          },
          "altText": {
            "type": "object",
            "description": "Альтернативный текст",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата создания",
            "example": "2025-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "projectId",
          "url",
          "type",
          "createdAt"
        ]
      },
      "ProjectLinkResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID ссылки",
            "example": "clx1234567890"
          },
          "url": {
            "type": "string",
            "description": "URL ссылки",
            "example": "https://github.com/user/project"
          },
          "title": {
            "type": "object",
            "description": "Название ссылки",
            "example": "GitHub Repository",
            "nullable": true
          },
          "projectId": {
            "type": "object",
            "description": "ID проекта",
            "example": "clx1234567890",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата создания",
            "example": "2025-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "url",
          "createdAt"
        ]
      },
      "ProjectCountsDto": {
        "type": "object",
        "properties": {
          "media": {
            "type": "number",
            "description": "Количество медиа файлов",
            "example": 3
          },
          "links": {
            "type": "number",
            "description": "Количество ссылок",
            "example": 2
          }
        },
        "required": [
          "media",
          "links"
        ]
      },
      "ProjectResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID проекта",
            "example": "clx1234567890"
          },
          "userId": {
            "type": "string",
            "description": "ID пользователя",
            "example": "clx1234567890"
          },
          "title": {
            "type": "string",
            "description": "Название проекта",
            "example": "E-commerce Platform"
          },
          "slug": {
            "type": "string",
            "description": "Slug проекта",
            "example": "e-commerce-platform"
          },
          "shortDesc": {
            "type": "object",
            "description": "Краткое описание",
            "nullable": true
          },
          "description": {
            "type": "object",
            "description": "Полное описание",
            "nullable": true
          },
          "techStack": {
            "description": "Технологический стек",
            "example": [
              "React",
              "Node.js",
              "PostgreSQL"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "visibility": {
            "type": "string",
            "description": "Видимость проекта",
            "enum": [
              "PUBLIC",
              "PRIVATE",
              "UNLISTED"
            ],
            "example": "PUBLIC"
          },
          "media": {
            "description": "Медиа файлы проекта",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectMediaResponseDto"
            }
          },
          "links": {
            "description": "Ссылки проекта",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectLinkResponseDto"
            }
          },
          "_count": {
            "description": "Количество элементов",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProjectCountsDto"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата создания",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата обновления",
            "example": "2025-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "userId",
          "title",
          "slug",
          "techStack",
          "visibility",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Название проекта",
            "example": "E-commerce Platform"
          },
          "slug": {
            "type": "string",
            "description": "Slug проекта (автоматически генерируется из названия, если не указан)",
            "example": "e-commerce-platform"
          },
          "shortDesc": {
            "type": "string",
            "description": "Краткое описание проекта",
            "example": "Современная платформа для электронной коммерции"
          },
          "description": {
            "type": "string",
            "description": "Полное описание проекта",
            "example": "Полнофункциональная платформа для электронной коммерции с поддержкой множества платежных систем"
          },
          "techStack": {
            "description": "Технологический стек проекта",
            "example": [
              "React",
              "Node.js",
              "PostgreSQL"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Название проекта",
            "example": "E-commerce Platform"
          },
          "shortDesc": {
            "type": "string",
            "description": "Краткое описание проекта",
            "example": "Современная платформа для электронной коммерции"
          },
          "description": {
            "type": "string",
            "description": "Полное описание проекта",
            "example": "Полнофункциональная платформа для электронной коммерции"
          },
          "techStack": {
            "description": "Технологический стек проекта",
            "example": [
              "React",
              "Node.js",
              "PostgreSQL"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "visibility": {
            "type": "string",
            "description": "Видимость проекта",
            "enum": [
              "PUBLIC",
              "PRIVATE",
              "UNLISTED"
            ],
            "example": "PUBLIC"
          }
        }
      },
...
    }
  }
}